feat: 도전과제 추가 및 리펙터링

+++ src/main/java/com/learnmore/legacy/domain/achievement/presentation/dto/response/AwardRes.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 @Builder
 public record AwardRes(
-        Long awardExp,
-        Long awardCredit,
+        Integer awardExp,
+        Integer awardCredit,
         List<AwardDto> achievementAward
 ) {
 }
